Split configuration and argument handling from main
