+++ FFXIV_TexTools_CLI/FFXIV_TexTools_CLI.csproj
@@ -55,6 +55,8 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Program.cs" />
+    <Compile Include="Configuration.cs" />
+    <Compile Include="Arguments.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
